Add viewer RPC backend for ironrdp-agent
