+++ crates/ironrdp-rpc/src/lib.rs
@@ -1,0 +1,5 @@
+//! Shared local RPC protocol and transport used by the agent and viewer.
+
+pub mod ipc;
+pub mod transport;
+pub mod wire;
