+++ crates/ironrdp-viewer/Cargo.toml
@@ -28,6 +28,7 @@ qoi = ["ironrdp/qoi"]
 qoiz = ["ironrdp/qoiz"]
 
 [dependencies]
+ironrdp-agent = { path = "../ironrdp-agent", version = "0.1" }
 ironrdp = { path = "../ironrdp", version = "0.17", features = ["connector", "cliprdr", "input", "pdu", "client", "client-all"] }
 ironrdp-cfg = { path = "../ironrdp-cfg", version = "0.1" }
 ironrdp-propertyset = { path = "../ironrdp-propertyset", version = "0.1" }
@@ -38,7 +39,7 @@ winit = { version = "0.30", features = ["rwh_06"] }
 softbuffer = "0.4"
 
 # CLI
-clap = { version = "4.6", features = ["derive", "cargo"] }
+clap = { version = "4.6", features = ["derive", "cargo", "env"] }
 inquire = "0.9"
 proc-exit = "2"
