+++ crates/ironrdp-viewer/src/cli.rs
@@ -62,6 +62,14 @@ struct Args {
     #[clap(short, long, value_parser)]
     log_file: Option<String>,
 
+    /// Run as a visible GUI host for the ironrdp-agent RPC protocol.
+    #[clap(long)]
+    rpc: bool,
+
+    /// Override the viewer RPC endpoint (defaults to the per-user viewer endpoint).
+    #[clap(long, requires = "rpc")]
+    rpc_endpoint: Option<String>,
+
     #[clap(long, value_parser)]
     gw_endpoint: Option<String>,
     #[clap(long, value_parser)]
@@ -70,6 +78,7 @@ struct Args {
     gw_pass: Option<String>,
 
     /// An address on which the client will connect.
+    #[clap(env = "RDP_HOSTNAME")]
     destination: Option<Destination>,
 
     /// Path to a .rdp file to read the configuration from.
@@ -77,7 +86,7 @@ struct Args {
     rdp_file: Option<PathBuf>,
 
     /// A target RDP server user name
-    #[clap(short, long)]
+    #[clap(short, long, env = "RDP_USERNAME")]
     username: Option<String>,
 
     /// An optional target RDP server domain name
@@ -85,7 +94,7 @@ struct Args {
     domain: Option<String>,
 
     /// A target RDP server user password
-    #[clap(short, long)]
+    #[clap(short, long, env = "RDP_PASSWORD")]
     password: Option<String>,
 
     /// Proxy URL to connect to for the RDCleanPath
@@ -225,6 +234,8 @@ pub struct ViewerConfig {
     // CLI-only settings that are not representable as `.rdp` file properties.
     log_file: Option<String>,
     dump_rdp: Option<PathBuf>,
+    rpc: bool,
+    rpc_endpoint: Option<String>,
 }
 
 impl ViewerConfig {
@@ -255,6 +266,19 @@ impl ViewerConfig {
         let log_file = args.log_file.clone();
         let dump_rdp = args.dump_rdp.clone();
 
+        if args.rpc {
+            if args.rdp_file.is_some() || dump_rdp.is_some() {
+                anyhow::bail!("--rpc cannot be combined with --rdp-file or --dump-rdp");
+            }
+            return Ok(Self {
+                builder: ConfigBuilder::new(),
+                log_file,
+                dump_rdp: None,
+                rpc: true,
+                rpc_endpoint: args.rpc_endpoint,
+            });
+        }
+
         // The library overlays everything expressible as a `.rdp` property: destination, credentials,
         // transport, channels, desktop size, audio, DVC proxies, etc.
         let builder = ConfigBuilder::from_property_set(&properties)?;
@@ -270,6 +294,8 @@ impl ViewerConfig {
             builder,
             log_file,
             dump_rdp,
+            rpc: false,
+            rpc_endpoint: None,
         })
     }
 
@@ -289,6 +315,14 @@ impl ViewerConfig {
     /// Path to dump the effective `.rdp` PropertySet to, if `--dump-rdp` was given.
     pub fn dump_rdp(&self) -> Option<&std::path::Path> {
         self.dump_rdp.as_deref()
+    }
+
+    pub fn rpc_mode(&self) -> bool {
+        self.rpc
+    }
+
+    pub fn rpc_endpoint(&self) -> Option<&str> {
+        self.rpc_endpoint.as_deref()
     }
 }
